Add files via upload
Hi guys. Here is the first part of the project written. I have put the variables in tables. It looks fine, but it takes a lot of space - especially the two last tables with variables. Perhaps when we narrow our variables down (if we do so), it won't take as lot of space.
If I just make one big table, it does not really change anything.